requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
openpyxl==3.1.2
numpy==1.24.3
Werkzeug==2.3.7
python-dotenv==1.0.0